refactored the tabs to new component
