+++ tailwind.config.js
@@ -108,6 +108,7 @@ module.exports = {
         900: "#1c2556",
         950: "#20253a",
         975: "#080B1A",
+        980: "#515151",
       },
     },
     extend: {
